KSENROLL-6559 - change "Break Colocation" warning dialog to a KRAD dialog rather than a browser dialog.
git-svn-id: https://svn.kuali.org/repos/student/enrollment/ks-enroll/trunk@52732 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-enroll-ui/src/main/java/org/kuali/student/enrollment/class2/courseoffering/helper/ActivityOfferingScheduleHelperImpl.java
@@ -339,7 +339,7 @@ public class ActivityOfferingScheduleHelperImpl implements ActivityOfferingSched
          * deleted all the RDLs and unchecked the colo checkbox. In this case, we should not be deleting the set,
          * just remove the AO from the coloset and save.)
          */
-        if (wrapper.getRequestedScheduleComponents().isEmpty()){
+        if (wrapper.getRequestedScheduleComponents().isEmpty() && !wrapper.isRemovedFromColoSet()){
             if (set != null && StringUtils.isNotBlank(set.getId())){
                 try {
                     StatusInfo statusInfo = getSchedulingService().deleteScheduleRequestSet(set.getId(),defaultContextInfo);
